+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Fixed
 * `crud.readview` resource cleanup on garbage collect (#379).
+* `doc/playground.lua` does not work with Tarantool 3 (#371).
 
 ## [1.3.0] - 27-09-23